Adding a run-action to upgrade Docker
